requirements.txt
confluent-kafka==2.6.0
psycopg2-binary==2.9.10
PyYAML==6.0.2
setuptools==75.1.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
wheel==0.44.0
